# webService/public/openapi/spec.yaml
openapi: 3.0.0
info:
  title: DataScope查询服务API
  description: 提供查询编辑、执行、版本控制及收藏功能的API接口
  version: 1.0.0
servers:
  - url: /api
    description: 默认API路径
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Query:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        dataSourceId:
          type: string
          format: uuid
        sqlContent:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
        queryType:
          type: string
          enum: [SQL, NATURAL_LANGUAGE]
        isFavorite:
          type: boolean
        executionCount:
          type: integer
        lastExecutedAt:
          type: string
          format: date-time
        tags:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
...
paths:
...
  /queries/favorites:
    get:
      tags:
        - 查询收藏
      summary: 获取用户收藏的查询列表
      description: 分页获取当前用户收藏的所有查询
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: 页码
        - in: query
          name: size
          required: false
          schema:
            type: integer
            default: 10
          description: 每页数量
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Query'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      size:
                        type: integer
                      totalPages:
                        type: integer
                  message:
                    type: string
        '401':
          description: 未授权
        '500':
          description: 服务器错误
  /queries/{queryId}/favorite:
    post:
      tags:
        - 查询收藏
      summary: 添加查询到收藏
      description: 将指定查询添加到用户的收藏列表
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queryId
          required: true
          schema:
            type: string
          description: 查询ID
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
        '401':
          description: 未授权
        '404':
          description: 查询不存在
        '500':
          description: 服务器错误
    delete:
      tags:
        - 查询收藏
      summary: 从收藏中移除查询
      description: 将指定查询从用户的收藏列表中移除
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queryId
          required: true
          schema:
            type: string
          description: 查询ID
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                  message:
                    type: string
        '401':
          description: 未授权
        '404':
          description: 查询不存在
        '500':
          description: 服务器错误
  /queries/{queryId}/favorite/status:
    get:
      tags:
        - 查询收藏
      summary: 检查查询是否被收藏
      description: 检查当前用户是否已收藏指定查询
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queryId
          required: true
          schema:
            type: string
          description: 查询ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      isFavorite:
                        type: boolean
                  message:
                    type: string
        '401':
          description: 未授权
        '404':
          description: 查询不存在
        '500':
          description: 服务器错误
...